feat: Add createFetch() action creator
Also export actionTypes with each action string

BREAKING CHANGE: Added promise to FetchAction['meta']
